Refactored AuthGuard & Implemented Token Validation
