+++ next-ts/app/(dashboard)/admin/(protected)/layout.tsx
@@ -1,4 +1,4 @@
-import WithAuth from '@/src/AuthGuard';
+import WithAuth from '@/src/lib/AuthGuard';
 import AdminLayout from '@/src/components/Admin/AdminLayout';
 
 export default function AdminProtectedLayout({
