+++ next-ts/tsconfig.json
@@ -29,6 +29,6 @@
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts",
     "**/*.mts"
-, "src/middleware"  ],
+, "src/middleware.ts"  ],
   "exclude": ["node_modules"]
 }
